refactor(core): consolidate schemas with Zod as single source of truth
Move Zod schemas to core package as the canonical source:
- Create zod-schemas.ts with all Zod schemas and derived types
- Update schema.ts to re-export from zod-schemas
- Update CLI output-schemas to import from core
- Remove duplicate ExecutionSummary interface from json.ts
- Make stashedId and restoredId required in Stash/Pop schemas
- Add stashedId/restoredId to stash/pop command output

This eliminates drift between TypeScript interfaces and Zod schemas
by deriving all types from Zod using z.infer<>.
