+++ packages/cli/src/schemas/output-schemas.ts
@@ -1,10 +1,8 @@
 /**
  * CLI output schemas using Zod.
  *
- * Defines the canonical JSON schema definitions for all CLI command outputs.
- * These schemas serve as the single source of truth for:
- * - Runtime validation in tests
- * - JSON Schema generation via `--schema` flag
+ * Re-exports schemas from @rundown-org/core (single source of truth) and
+ * defines CLI-specific schemas and command mappings.
  *
  * @module schemas/output-schemas
  */
@@ -12,141 +10,73 @@
 import { z } from 'zod';
 
 // ============================================================================
-// Error Codes
+// Re-export from Core (Single Source of Truth)
 // ============================================================================
 
-export const ErrorCodeSchema = z.enum([
-  'NO_ACTIVE_RUNBOOK',
-  'RUNBOOK_NOT_FOUND',
-  'STEP_NOT_FOUND',
-  'INVALID_SYNTAX',
-  'VALIDATION_ERROR',
-  'NO_STASHED_RUNBOOK',
-  'AGENT_BINDING_ERROR',
-  'SCENARIO_NOT_FOUND',
-  'FILE_ERROR',
-  'UNKNOWN_ERROR',
-]).describe('Error code identifying the type of error that occurred');
+export {
+  // Error codes
+  ErrorCodeSchema,
+
+  // Shared schemas
+  PositionSchema,
+  RunbookContextSchema,
+  ErrorDetailsSchema,
+
+  // Response schemas
+  ErrorResponseSchema,
+  ActionResponseSchema,
+  StatusResponseSchema,
+  CheckResponseSchema,
+  EchoResponseSchema,
+  StashResponseSchema,
+  PopResponseSchema,
+
+  // List schemas
+  ActiveRunbookEntrySchema,
+  AvailableRunbookEntrySchema,
+  ActiveRunbookListSchema,
+  AvailableRunbooksListSchema,
+
+  // Check schemas
+  CheckValidationErrorSchema,
+  RunbookStatsSchema,
+
+  // Scenario schemas
+  ScenarioEntrySchema,
+  ScenarioDetailSchema,
+  ScenarioListSchema,
+  ScenarioRunResponseSchema,
+  ScenarioErrorResponseSchema,
+
+  // Prune schemas
+  PruneEntrySchema,
+  PruneResponseSchema,
+
+  // Execution schemas
+  ExecutionSummarySchema,
+  StepQueuedResponseSchema,
+  AgentBoundResponseSchema,
+  RunCommandResponseSchema,
+} from '@rundown-org/core';
 
 // ============================================================================
-// Shared Types
-// ============================================================================
-
-export const PositionSchema = z.object({
-  current: z.string().describe('Current step number or identifier'),
-  total: z.union([z.number(), z.string()]).describe('Total number of steps or status indicator'),
-  substep: z.string().optional().describe('Optional substep identifier'),
-}).describe('Current position within the runbook execution');
-
-export const RunbookContextSchema = z.object({
-  file: z.string().describe('Path to the runbook file'),
-  state: z.string().describe('Current runbook state or status'),
-  prompted: z.boolean().optional().describe('Whether the runbook is awaiting user input'),
-}).describe('Context information about the active runbook');
-
-export const ErrorDetailsSchema = z
-  .object({
-    requested: z.string().optional().describe('The item or resource that was requested'),
-    available: z.array(z.string()).optional().describe('List of available options'),
-    suggestion: z.string().optional().describe('Suggested resolution or alternative'),
-    path: z.string().optional().describe('File path related to the error'),
-    searchedLocations: z.array(z.string()).optional().describe('Locations that were searched'),
-    line: z.number().optional().describe('Line number where the error occurred'),
-  })
-  .describe('Additional details about an error')
-  .passthrough();
-
-// ============================================================================
-// Response Schemas
+// CLI-Specific Schemas
 // ============================================================================
 
 /**
- * Error response schema.
+ * Prompt response schema.
  *
- * All error responses must have result=false and an error message.
+ * Simple output wrapper for prompt command (CLI-only, not in core).
  */
-export const ErrorResponseSchema = z
-  .object({
-    result: z.literal(false).describe('Always false for error responses'),
-    error: z.string().describe('Error message describing what went wrong'),
-    code: ErrorCodeSchema.optional().describe('Error code for programmatic handling'),
-    details: ErrorDetailsSchema.optional().describe('Additional error context'),
-  })
-  .describe('Error response indicating command execution failure')
-  .passthrough();
+export const PromptResponseSchema = z.object({
+  output: z.string().describe('Prompt output text'),
+}).describe('Response from the prompt command');
 
 /**
- * Action response schema (pass, fail, goto, stop, complete).
+ * Runbook schema - unified schema for ls output with status field.
  *
- * Action responses include the action performed and position changes.
- * Uses `result` boolean to indicate action success (PASS = true, FAIL = false).
- */
-export const ActionResponseSchema = z
-  .object({
-    action: z.string().describe('Type of action performed'),
-    command: z.string().optional().describe('Command executed for this action'),
-    from: PositionSchema.optional().describe('Starting position before action'),
-    to: PositionSchema.optional().describe('Position after action execution'),
-    complete: z.boolean().optional().describe('Whether the runbook completed'),
-    stopped: z.boolean().optional().describe('Whether the runbook was stopped'),
-    runbook: RunbookContextSchema.optional().describe('Runbook context information'),
-    // Flat format fields
-    file: z.string().optional().describe('Path to the runbook file'),
-    state: z.string().optional().describe('Runbook state after action'),
-    prompted: z.boolean().optional().describe('Whether awaiting user input'),
-    result: z.boolean().optional().describe('Whether the action succeeded'),
-    message: z.string().optional().describe('Status message from the action'),
-    position: PositionSchema.optional().describe('Current position after action'),
-  })
-  .describe('Response from a step action command')
-  .passthrough();
-
-/**
- * Status response schema.
- *
- * Uses flat structure per CLI-OUTPUT-SPEC:
- * - `file`/`state`/`prompted` at top level (not nested in `runbook`)
- * - `position` for step position (current/total/substep)
- * - `step` for step details (name/description)
- */
-export const StatusResponseSchema = z
-  .object({
-    active: z.boolean().describe('Whether a runbook is currently active'),
-    stashed: z.boolean().describe('Whether a runbook is stashed'),
-    // Flat structure fields
-    file: z.string().optional().describe('Path to the active runbook file'),
-    state: z.string().optional().describe('Current runbook execution state'),
-    prompted: z.boolean().optional().describe('Whether awaiting user input'),
-    position: PositionSchema.optional().describe('Current step position'),
-    step: z
-      .object({
-        name: z.string().describe('Step name or identifier'),
-        description: z.string().optional().describe('Step description'),
-      })
-      .optional()
-      .describe('Current step details'),
-    lastAction: z
-      .object({
-        action: z.string().describe('Last action performed'),
-        result: z.boolean().optional().describe('Result of the last action'),
-      })
-      .optional()
-      .describe('Last action information'),
-  })
-  .describe('Response from the status command')
-  .passthrough();
-
-/**
- * Runbook schema - unified schema for ls and prune output.
- *
- * Both commands list runbook entries with common fields:
- * - id: state file identifier
- * - runbook: runbook filename
- * - status: state status (active, stashed, completed, stale, orphaned)
- *
- * Optional fields vary by command:
- * - step/total: position info (ls shows current position)
- * - title: when available from runbook metadata
+ * This is a CLI-specific variant that includes status field for display.
+ * Maps to ActiveRunbookEntrySchema from core but with status required.
  */
 export const RunbookSchema = z.object({
   id: z.string().describe('Unique state file identifier'),
@@ -158,238 +88,42 @@ export const RunbookSchema = z.object({
 }).describe('Runbook state entry');
 
 /**
- * Available runbook entry (ls --all output).
- *
- * Different from RunbookStateEntry - this lists runbook files,
- * not active state entries.
- */
-export const AvailableRunbookEntrySchema = z.object({
-  name: z.string().describe('Runbook name from metadata or filename'),
-  description: z.string().optional().describe('Runbook description from metadata'),
-  path: z.string().describe('File path to the runbook'),
-}).describe('Available runbook file entry');
-
-/**
- * List of runbooks (ls or prune output).
+ * List of runbooks (ls output with status).
  */
 export const RunbookListSchema = z.array(RunbookSchema).describe('List of runbook state entries');
 
 /**
- * List of available runbooks (ls --all output).
+ * Validation error entry (alias for consistency with CLI naming).
  */
-export const AvailableRunbooksListSchema = z.array(AvailableRunbookEntrySchema).describe('List of available runbook files');
+export { CheckValidationErrorSchema as ValidationErrorSchema } from '@rundown-org/core';
 
 /**
- * Validation error entry.
+ * Scenario schema (alias for CLI naming consistency).
  */
-export const ValidationErrorSchema = z.object({
-  message: z.string().describe('Error message'),
-  line: z.number().optional().describe('Line number where error occurred'),
-}).describe('Validation error entry');
+export { ScenarioEntrySchema as ScenarioSchema } from '@rundown-org/core';
 
 /**
- * Check response schema.
+ * Scenario show response (alias for CLI naming consistency).
  */
-export const CheckResponseSchema = z.object({
-  valid: z.boolean().describe('Whether the runbook is valid'),
-  errors: z.array(ValidationErrorSchema).describe('List of validation errors'),
-  stats: z
-    .object({
-      steps: z.number().describe('Total number of steps'),
-      substeps: z.number().describe('Total number of substeps'),
-    })
-    .optional()
-    .describe('Runbook statistics'),
-}).describe('Response from the check command');
-
-/**
- * Scenario schema (scenario ls output).
- */
-export const ScenarioSchema = z.object({
-  name: z.string().describe('Scenario name'),
-  expected: z.string().describe('Expected scenario outcome'),
-  description: z.string().optional().describe('Scenario description'),
-  tags: z.string().optional().describe('Comma-separated scenario tags'),
-}).describe('Scenario definition');
-
-/**
- * Scenario list.
- */
-export const ScenarioListSchema = z.array(ScenarioSchema).describe('List of scenarios in a runbook');
-
-/**
- * Scenario show response (extends schema with commands).
- */
-export const ScenarioShowResponseSchema = ScenarioSchema.extend({
-  commands: z.array(z.string()).optional().describe('List of commands in the scenario'),
-}).describe('Response from scenario show command');
-
-/**
- * Scenario run response.
- *
- * Uses `passed` to indicate scenario outcome (not `result` - scenario verification, not workflow).
- */
-export const ScenarioRunResponseSchema = z.object({
-  scenario: z.string().describe('Scenario name'),
-  expected: z.string().describe('Expected outcome'),
-  actual: z.string().describe('Actual outcome'),
-  passed: z.boolean().describe('Whether the scenario passed'),
-  message: z.string().optional().describe('Additional status message'),
-}).describe('Response from scenario run command');
-
-/**
- * Scenario error response.
- */
-export const ScenarioErrorResponseSchema = z.object({
-  error: z.literal(true).describe('Always true for error responses'),
-  message: z.string().describe('Error message'),
-  available: z.array(z.string()).optional().describe('Available scenarios'),
-}).describe('Scenario error response');
-
-/**
- * Echo response schema.
- *
- * Uses `output` for echoed text and `result` boolean for success status.
- */
-export const EchoResponseSchema = z.object({
-  result: z.boolean().describe('Whether the echo command succeeded'),
-  output: z.string().optional().describe('Echoed output text'),
-  error: z.string().optional().describe('Error message if command failed'),
-  exitCode: z.number().optional().describe('Exit code of the echo command'),
-}).describe('Response from the echo command');
-
-/**
- * Prompt response schema.
- *
- * Simple output wrapper for prompt command.
- */
-export const PromptResponseSchema = z.object({
-  output: z.string().describe('Prompt output text'),
-}).describe('Response from the prompt command');
-
-/**
- * Stash response schema.
- *
- * Uses action='stash' (present tense verb).
- */
-export const StashResponseSchema = z
-  .object({
-    result: z.boolean().describe('Whether the stash operation succeeded'),
-    action: z.literal('stash').describe('Action type'),
-    stashedId: z.string().optional().describe('ID of the stashed runbook'),
-    runbook: RunbookContextSchema.optional().describe('Runbook context'),
-    file: z.string().optional().describe('Path to the runbook file'),
-    state: z.string().optional().describe('Runbook state'),
-    message: z.string().optional().describe('Status message'),
-    position: PositionSchema.optional().describe('Position when stashed'),
-  })
-  .describe('Response from the stash command')
-  .passthrough();
-
-/**
- * Pop response schema.
- *
- * Uses action='pop'.
- */
-export const PopResponseSchema = z
-  .object({
-    result: z.boolean().describe('Whether the pop operation succeeded'),
-    action: z.literal('pop').describe('Action type'),
-    restoredId: z.string().optional().describe('ID of the restored runbook'),
-    runbook: RunbookContextSchema.optional().describe('Runbook context'),
-    file: z.string().optional().describe('Path to the runbook file'),
-    state: z.string().optional().describe('Runbook state'),
-    message: z.string().optional().describe('Status message'),
-    position: PositionSchema.optional().describe('Position when restored'),
-    step: z
-      .object({
-        name: z.string().optional().describe('Step name'),
-        description: z.string().optional().describe('Step description'),
-        prompted: z.boolean().optional().describe('Whether waiting for input'),
-      })
-      .optional()
-      .describe('Current step details'),
-  })
-  .describe('Response from the pop command')
-  .passthrough();
-
-/**
- * Execution summary schema.
- *
- * Output from commands that use JSONSubscriber.getSummary() like goto.
- */
-export const ExecutionSummarySchema = z
-  .object({
-    runbookId: z.string().optional().describe('Runbook state identifier'),
-    runbook: z.string().optional().describe('Runbook filename'),
-    status: z.enum(['complete', 'stopped', 'running']).describe('Execution status'),
-    stepsExecuted: z.number().describe('Number of steps executed'),
-    commandsRun: z.number().describe('Number of commands run'),
-    commandsFailed: z.number().describe('Number of commands that failed'),
-    finalPosition: z
-      .object({
-        current: z.string().describe('Final step number'),
-        total: z.union([z.number(), z.string()]).describe('Total steps'),
-        substep: z.string().optional().describe('Final substep'),
-      })
-      .optional()
-      .describe('Final execution position'),
-    message: z.string().optional().describe('Execution message'),
-    events: z
-      .array(
-        z
-          .object({
-            type: z.string().describe('Event type'),
-            timestamp: z.string().optional().describe('Event timestamp'),
-          })
-          .describe('Execution event')
-          .passthrough()
-      )
-      .optional()
-      .describe('Execution events'),
-  })
-  .describe('Response from commands with execution summary')
-  .passthrough();
-
-/**
- * Step queued response schema (run --step output).
- *
- * Emitted when a step is queued for execution via `run --step`.
- */
-export const StepQueuedResponseSchema = z.object({
-  action: z.literal('step_queued').describe('Action type for step queue'),
-  stepId: z.string().describe('Step identifier that was queued'),
-  runbook: z.string().optional().describe('Runbook filename'),
-});
-
-/**
- * Agent bound response schema (run --agent output).
- *
- * Emitted when an agent is bound to a step via `run --agent`.
- */
-export const AgentBoundResponseSchema = z.object({
-  action: z.literal('agent_bound').describe('Action type for agent binding'),
-  agent: z.string().describe('Agent identifier that was bound'),
-  stepId: z.string().describe('Step identifier the agent is bound to'),
-});
-
-/**
- * Combined run command response schema (all variants).
- *
- * The run command can produce different output formats depending on flags:
- * - Default: ExecutionSummarySchema (full execution)
- * - --step: StepQueuedResponseSchema (step queue confirmation)
- * - --agent: AgentBoundResponseSchema (agent binding confirmation)
- */
-export const RunCommandResponseSchema = z.union([
-  ExecutionSummarySchema,
-  StepQueuedResponseSchema,
-  AgentBoundResponseSchema,
-]);
+export { ScenarioDetailSchema as ScenarioShowResponseSchema } from '@rundown-org/core';
 
 // ============================================================================
 // Command to Schema Mapping
 // ============================================================================
+
+import {
+  StatusResponseSchema,
+  ActionResponseSchema,
+  CheckResponseSchema,
+  EchoResponseSchema,
+  StashResponseSchema,
+  PopResponseSchema,
+  ScenarioListSchema,
+  ScenarioDetailSchema,
+  ScenarioRunResponseSchema,
+  ExecutionSummarySchema,
+  AvailableRunbooksListSchema,
+} from '@rundown-org/core';
 
 /**
  * Maps CLI command names to their output schemas.
@@ -409,10 +143,10 @@ export const COMMAND_SCHEMAS: Record<string, z.ZodSchema> = {
   check: CheckResponseSchema,
   echo: EchoResponseSchema,
   prompt: PromptResponseSchema,
-  run: RunCommandResponseSchema,
+  run: z.union([ExecutionSummarySchema, z.object({ action: z.literal('step_queued'), stepId: z.string(), runbook: z.string().optional() }), z.object({ action: z.literal('agent_bound'), agent: z.string(), stepId: z.string() })]),
   ls: z.union([RunbookListSchema, AvailableRunbooksListSchema]),
   prune: RunbookListSchema,
   'scenario ls': ScenarioListSchema,
-  'scenario show': ScenarioShowResponseSchema,
+  'scenario show': ScenarioDetailSchema,
   'scenario run': ScenarioRunResponseSchema,
 };
